First pass at performing a simple method-level liveness analysis to remove unnecessary method calls
In order to fix LValue graph traversal, we now explicitly model `merges` in the LValue graph, rather than trying to (incorrectly) stuff them into `upstream`.

For now we simply stub out the method call with an appropriate number of POPs followed by a *CONST. We do not try to clean up the resultant bytecode in any way. A pure-cleanup pass to deal with all the stubbed-out bytecodes seems to be the way forward

We also return method parameter liveness information, though do not yet make use of it

The old liveness code has been removed.