Fix `loopReturn1` test with more aggressive `SSA.Phi#checkLinks` implementation
We should always perform the `BytecodeToSSA#merge`, even if the two items being merged are the same, because it is perfectly possible for a Phi value to be one of its own incomings.

Also simplify `merge` somewhat since the `v1` will always be a `SSA.Phi`

For some reason `joptimize.MainTests.simple.Sudoku.run` now fails when run together with the rest of the suite, but passes when run alone. Fixing that TBD