Re-implement purity analysis, get ConstantMethod folding tests passing
Currently purity inference is interleaved together with type inference. This is not strictly necessary, as purity inference can be made a separate pass after type inference completes, but this way is easier to implement as it can share the same call-graph-traversal logic that type inference uses.

`ITypeLattice` and a new `PurityLattice` are now packaged together into `CombinedLattice`, which does the inference of both types and purity concurrently with minimal changes to the `OptimisticAnalyze` logic. Some additional boilerplate is now necessary around these APIs to pack and unpack the combined `(IType, purity: Boolean)` tuples