Fix `Inheritance.lambdaSpecialized` by mangling specialized method names using original return type.
The JVM can rely on the return type of a method to perform overload resolution. Since we change the return type during specialization, this can cause clashes when multiple return-type-overloaded methods are specialized to have the same return type. The solution is to include the original return type in the name mangling to disambiguate