Get last `Chaining.hello8` test passing
Still not optimized, but at least it runs

We needed to narrow the generation of `InvokeVirtual`/`InvokeInterface` bytecodes to use the inferred receiver type, rather than the original receiver type. This makes it match ProgramAnalyzer, which walks methods based on the inferred receiver type

In order to support this, we needed to also allow conversion from `InvokeInterface` into `InvokeVirtual`, when the interface type is narrowed into a class type. We do this by collapsing`InvokeInterface` into `InvokeVirtual` and adding a `interface` flag, making it easy to modify without re-wiring the SSA graph.

Now that we are narrowing callsite receivers, we also need to be able to handle the case where a callsite is made against a class but the called method is defined in the superclass. This is done by modifying `resolvePossibleSigs`, and making `BytecodeDCE` use that instead of naive `findSubtypes` checks