Split up `ClassManager` into `Frozen` and `Dynamic` versions
This lets us clearly distinguish the parts of the code that need to load classes (e.g. lazily in `ProgramAnalyzer`) and those that shouldn't (all of Backend)

Also created a new `ClassManager.Dynamic` for `BytecodeDCE`, since by the time it gets to that point all the specialization and inlining means the classes and methods loaded earlier during `ProgramAnalyzer` will be invalid.

10 tests still failing...